Refactored tilt correction and tilt solution tests.
